app3/eyecare_backend/requirements.txt
Flask==3.1.2
flask-cors==6.0.1
Flask-Mail==0.10.0
Flask-Limiter==3.8.0
flasgger==0.9.7.1
PyMySQL==1.1.2
psycopg2-binary==2.9.9
numpy==2.3.5
pandas==2.3.3
scikit-learn==1.7.2
lightgbm==4.6.0
imbalanced-learn==0.14.0
python-dotenv==1.2.1
gunicorn==23.0.0
pytest==8.3.5
pytest-cov==6.0.0
redis==5.0.8
python-dateutil==2.9.0
sentry-sdk[flask]==2.22.0
zeroconf==0.148.0

